fixed pinging (commit retry)
the master would get no response from ping (a change made to fix the cutoff of the result that could happen because of the ping response). fixed by increasing the size and adding a blank " " as a response to not mess with the response. (last commit failed hence the retry)